requirements.txt
asgiref==3.2.10
backcall==0.2.0
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.1
django-heroku==0.3.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
flake8==3.8.4
gunicorn==20.0.4
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
mccabe==0.6.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
traitlets==5.0.4
wcwidth==0.2.5
whitenoise==5.2.0
python-decouple==3.3
Pillow==8.0.0
